Update quick-start: token and hotkeys, more detail
Update quick-start by fleshing out with more detail. Of particular
importance is describing where to find the Authorization token in
Joplin Desktop and explaining away the clobbering of `Ctrl+J` system
keybinding.

Broke up text in a way that gives the impression of good document
formatting despite the constraints of markdown.